[PHASE-10.2] ORGANIZE: Standardize report naming and complete Phase 10.2.2 deep analysis
Why: Improve discoverability and organization of modernization documentation
Who: Claude + User
Phase: 10.2 - Deep Refactoring Assessment

## Report File Reorganization

Standardized all phase report filenames with consistent zero-padded numbering:
- Phase 2-9 reports: PHASE_0X format (e.g., PHASE_02, PHASE_08)
- Phase 10 subphase reports: PHASE_10.XX.YY format (e.g., PHASE_10.02.02)
- Deleted obsolete intermediate reports (char_audit_raw.txt, etc.)

Benefits:
- Files sort alphabetically by phase order
- Clear hierarchy: Phase → Subphase → Document type
- Easier navigation in file listings
- Consistent pattern across all documentation

## Phase 10.2.2 Deep Analysis (NEW)

Completed systematic deep refactoring analysis framework:

### New Analysis Reports (7 files):
- PHASE_10.02.02_refactor_report_on_m2alloc.c.md
- PHASE_10.02.02_refactor_report_on_cexecute.c.md
- PHASE_10.02.02_refactor_report_on_admin.c.md
- PHASE_10.02.02_refactor_report_on_update.c.md
- PHASE_10.02.02_refactor_report_on_forms.c.md
- PHASE_10.02.02_refactor_report_on_navy.c.md
- PHASE_10.02.02_refactor_report_on_combat.c.md

### Architectural Debt Supplements (5 files):
- PHASE_10.02.02_ARCHITECTURAL_DEBT_admin.c.md
- PHASE_10.02.02_ARCHITECTURAL_DEBT_update.c.md
- PHASE_10.02.02_ARCHITECTURAL_DEBT_forms.c.md
- PHASE_10.02.02_ARCHITECTURAL_DEBT_navy.c.md
- PHASE_10.02.02_ARCHITECTURAL_DEBT_combat.c.md

### Project Management:
- PHASE_10.02.02_deep_analysis_todo.md - Master TODO tracking 30 files
- PHASE_10.02.02_CONSTANT_INVENTORY.md - Magic number catalog

### Key Findings:

**Three File Type Categories Identified:**
1. Pure Utilities (m2alloc.c) - ⭐⭐⭐⭐⭐ Zero coupling, perfect testability
2. Core Game Engines (cexecute.c) - ⭐⭐⭐ Necessary coupling, acceptable
3. Mixed Responsibility (admin.c) - ⭐⭐ Misplaced functions, needs reorganization

**Analysis Completed (3 of 30 files):**
- m2alloc.c: Gold standard, no changes needed
- cexecute.c: Core engine, config cleanup only (8 magic numbers)
- admin.c: 29% misplaced functions, 50+ magic numbers

**Remaining**: 27 files in user-specified analysis order

## Session Memory Checkpoints (11 files)

Added comprehensive session continuity documentation:
- Phase 10.1 completion checkpoints (5 files)
- Phase 10.2.1 completion checkpoints (2 files)
- Phase 10.2.2 analysis checkpoints (4 files)

Enables seamless continuation across sessions with full context preservation.

## Source Code Updates (Minimal)

Minor formatting improvements from Phase 10.2.1:
- cexecute.c: Formatting consistency
- extcmds.c: Formatting consistency
- forms.c: Formatting consistency

## Testing Infrastructure

Updated test_warnings.sh with enhanced baseline tracking for Phase 10.2.2 analysis validation.

## Files Changed Summary

Renamed: 47 report files (consistent naming)
Deleted: 4 obsolete intermediate files
New: 23 analysis and session memory files
Modified: 4 source/script files (minimal formatting)

Progress: Phase 10.2.2 analysis 10% complete (3 of 30 files)
Next: Continue deep analysis starting with check.c

+++ forms.c
@@ -824,12 +824,12 @@ void change(void) {
             mvprintw(5, 0, "active is %d", curntn->active);
         mvprintw(6, 0, "tax_rate...... %2d%%", curntn->tax_rate);
         mvprintw(7, 0, "inflation..... %2d%%", curntn->inflation);
-        i = (int)10 * curntn->tax_rate - curntn->popularity - curntn->terror
+        i = 10 * curntn->tax_rate - curntn->popularity - curntn->terror
             - 3 * curntn->charity;
         if (i < 0)
             i = 0;
         mvprintw(8, 0, "peasant revolt %2d%%", (i * PREVOLT) / 100);
-        i = (int)5 * curntn->tax_rate - curntn->prestige;
+        i = 5 * curntn->tax_rate - curntn->prestige;
         if (i < 0)
             i = 0;
         mvprintw(9, 0, "other revolt.. %2d%%", (i * PREVOLT) / 100);
